feat: use photon classifier within fit algorithms
feat: use the GSFC YAPC k-nearest neighbors algorithm to determine photon event weights

refactor: move functions for ATL03 fits into separate program

ci: rename automatic time update PR title
