+++ .github/workflows/auto-update-files.yml
@@ -52,4 +52,4 @@ jobs:
       uses: peter-evans/create-pull-request@v3
       with:
         assignees: ${{ github.actor }}
-        title: "Automatic time updates"
+        title: "data: automatic time updates"
