+++ environment.yml
@@ -4,6 +4,7 @@ channels:
 dependencies:
   - python>=3.6
   - notebook
+  - cython
   - numpy
   - scipy
   - scikit-learn
