+++ icesat2_toolkit/__init__.py
@@ -14,6 +14,8 @@ ICESat-2 mission and to support its science applications
 
 Documentation is available at https://read-icesat-2.readthedocs.io
 """
+import icesat2_toolkit._fit
+import icesat2_toolkit.fit
 import icesat2_toolkit.time
 import icesat2_toolkit.utilities
 from icesat2_toolkit.convert import convert
