+++ requirements-dev.txt
@@ -2,3 +2,4 @@ flake8
 pytest>=4.6
 pytest-cov
 numpy
+cython
