+++ requirements.txt
@@ -1,3 +1,4 @@
+cython
 numpy
 scipy
 mpi4py
